Recent serie (#37)
* I offer this to you in the hope that you like it.
A note is added to the cover and the title of the series or film.

* Added the ability to change the theme in the application settings.

* I added a <Trailer> button to the Series and Movies page.
If a trailer exists in the json string, it opens in YouTube; otherwise, a search for the trailer in YouTube is performed.

There's also a fix for app_<code_language>.dart because some variables had uppercase letters.

* Allows you to have an indicator for recent episodes (- 15 days)

* Allows you to have an indicator for recent series (- 15 days)

---------

Co-authored-by: saezolivier6633063-alt <saez.olivier@hotmail.fr>